pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-scrape"
version = "1.0.0"
description = "Python web scraping automation with Playwright and Excel integration"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Auto Scrape Team"}
]

dependencies = [
    "playwright>=1.40.0",
    "undetected-playwright-python>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "aiofiles>=23.2.1",
    "asyncio-throttle>=1.0.2",
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.1.9",
    "PyYAML>=6.0.1",
    "jinja2>=3.1.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"